% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bf.R
\name{get_all_previsions_available}
\alias{get_all_previsions_available}
\title{Get all data available at the current date}
\usage{
get_all_previsions_available()
}
\value{
a tibble
}
\description{
bison fute website uses data from past date and future date
for its previsions color of departement. See http://www.bison-fute.gouv.fr/previsions,10564.html
}
